requirements_base.txt
numpy>=1.23
pandas>=2.0
Pillow>=9.5
tqdm>=4.66
scikit-learn>=1.3
matplotlib>=3.7
openpyxl>=3.1